Remove zsh noclobber, suggest ignoring rc files
In case the user has some zsh RC files (e.g. .zshrc), it should be better to ignore them when running macos-guest-virtualbox.sh, by adding -f option to zsh's command line. Otherwise, the script might not run smoothly (e.g. rm might ask for confirmation of every single file deletion if it was alias to "rm -i" in .zshrc), or even break (if the noclobber option of zsh was set up in .zshrc).
As an additional caution, the option noclobber can be deactivated in the beginning of the script --- that way, lines overwriting or adding to .sh or .txt files will still work even though .zshrc was read and set this option.